+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Beaker.workspace.datasets()`.
 - Added `Beaker.workspace.experiments()`.
 - Added `Beaker.workspace.images()`.
+- Added `Beaker.workspace.list()`.
 
 ### Removed